requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.8.0
resend>=2.0.0

# Database
aiomysql>=0.2.0
mysql-connector-python>=8.0.0

# Audio processing
pydub>=0.25.1

# Web framework (for webhooks)
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
httpx>=0.27.0
tenacity>=8.2.0
